+++ QuantConnect.Lean.sln
@@ -56,7 +56,7 @@ Project("{888888A0-9F3D-457C-B088-3A5042F75D52}") = "QuantConnect.Algorithm.Pyth
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "QuantConnect.Algorithm.Framework", "Algorithm.Framework\QuantConnect.Algorithm.Framework.csproj", "{75981418-7246-4B91-B136-482728E02901}"
 EndProject
-Project("{888888A0-9F3D-457C-B088-3A5042F75D52}") = "QuantConnect.ApiPython", "ApiPython\QuantConnect.ApiPython.pyproj", "{1C47F4DB-2AFF-4FAE-9142-B33BE654A516}"
+Project("{888888A0-9F3D-457C-B088-3A5042F75D52}") = "QuantConnect.PythonToolbox", "PythonToolbox\QuantConnect.PythonToolbox.pyproj", "{1C47F4DB-2AFF-4FAE-9142-B33BE654A516}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
